Prevent timeout events from being emitted when item is already fetched
Fixes #136. It was possible for timeout events to be emitted after
a page had already emitted a 404 event. The reason for this is
likely the fact that when 404's are handled, the HTTP response
socket is closed, but the request socket isn't. I'm not too familiar
with the node http api, but I'm guessing this is why the timeout
event for that request can still be emitted.