Added decodeResponses option
With this enabled, the crawler looks for a character encoding in
the response headers and falls back to the first 512 bytes of the
response body, where it looks for a <meta> tag that specifies a
charset. This information is then passed to iconv-lite that does
the actual decoding.

By default this option is disabled.

I also added a test.