I saw a process.nextTick right after crawler.emit("fetchstart"), I think yeah this gives user time to change the request options. Then I think wait isn't `.emit` and `.on` are all synchronoous? So I don't understand why do we need the process.nextTick() after fetchstart.
In the start() function of crawler.js I see in the end there's a process.nextTick, and that is used for collecting user custom options before we start even if we call the start() before those custom options.
But the crawler.interval cannot get picked up this way. Maybe I'm wrong but should the run loop(setTimeout part) inside nextTick aswell?

The resource here was decleared but never used in the function, this
leads ```js
if (resources.reduce(function(prev, current) {
    return prev || current === URL;
}, false)) {
    return list;
}```  malfunctioning? I change it with list so that this function
can work again.