Always close HTTP response streams on "end"
There's a recurring issue where some requests are left hanging,
they're in a state where they're accepting data, but they're never
finished, and they cause the crawler to hang indefinitely.

This is an attempt at fixing that by always closing HTTP responses
on their "end" event.

The only hole in the previous logic that I can see is if the
complete property for some reason isn't always set on the
response when it is in fact complete. I don't see why that would
be, but on the other hand that property isn't documented, so who
knows...