Explicitly get only protocol, host and port for the touchedHosts list
Previously we constructed the touchedHost from a full URL simply
by stripping the pathname. This left us with cases where the query
string was still intact which broke the robots-parser module. We
want to be explicit, not implicit, about what we're looking for in
the host variable. This fixes that.