Ensure response sockets are destroyed on redirect and 404
In #201, we ensured response sockets would be destroyed when data
was recieved normally or when we got a total error. As reported
in #229, crawls still hang, and seem to do in response to 404/410
responses. We didn't previously call the destroy method on the
response socket when we got such responses (only the end method),
but this might solve the hang issue.