processURL method now constructs full queueItem
This means we have a single place where we manufacture queueItems,
and the previous parsedURL template is no longer used. This is
good because parsedURL was basically just an incomplete queueItem.
This change also reassigns the responsibility of creating
queueItems from the queue.add method to crawler.processURL. This
simplifies the queue.add method without making crawler.processURL
more complex.