Added force argument to queue.add method
When truthy, it will make sure the passed queue item is always
added to the queue, even if the URL it represents has already been
added previously.

However, because queueItems are directly mutable, we also ensure
that duplicate queueItems can't be added, only queueItems
representing the same URL are allowed.