Constructor function now takes a single argument: initialURL
All the URL properties we need can (naturally) be infered from
this single argument. And since URL's more often appear as an
entire string than as individual parts, it makes a lot of sense
to supply an entire URL to the constructor.

This also helps the issue where previously the crawler didn't take
an argument for initialProtocol, which in many cases would be more
relevant than initialPort. The initial port would most often be
the default one for the initial protocol, but the developer had to
set those properties in a reverse order. When an entire URL is
supplied, it becomes very natural to always supply the protocol,
and only optionally supply the port.