Added the concept of download conditions
This new API works in much the same way as fetch conditions, but
every callback that's registered is evaluated once the headers for
a resource have been fetched. This means the developer has access
to the Content-Type and Content-Length etc.

This also means that we're removing the dreaded
fetchWhitelistedMimeTypesBelowMaxDepth option AND in turn changing
how the maxDepth option behaves. These changes should hopefully
make sense to developers that now have very flexible ways of
determining which resources to fetch and how.