Only update queue items asynchronously
Fixes #257

The final puzzle piece that enables asynchronous queue
implementations! It adds an `update` method to the queue that
takes a URL, an object that describes the updates to be made to
the queue item, and a callback.

I've also made sure that all the places where the crawler mutates
queueItems that are *in the queue* now use this method.

And of course, I've added some tests.