Fixed a race condition in crawl loop
As reported by @HugoMFernandes in #298, a race condition can be
triggered in the crawler, where a short main interval combined
with a slow server response makes the crawler believe that no
queueItem is being processed, and it's time to emit a complete
event. The cause of this was that queue items were being marked
as fetched before their response body had been successfully
unzipped (which is an async operation).

Fixes #298