Added documentation for queue module
Also updated its functionality to some degree. Firstly, I made
members that are meant to be private non-enumerable and renamed
them to start with an underscore. I also changed the defrost
method to use a classical for loop rather than a for..in loop.
This is possible due to the fact that the length property is
included in the frozen file, and the property names of the frozen
queue objects are simply numbers.