Tolerate queue duplicate errors when starting crawler
If we stop the crawler and then start it again, we would always
encounter this error. With this change, that's no longer the case.
But we also don't want to add the initial URL to the queue only if
the queue is empty, since users might want to populate the queue
before starting the crawler for the first time, but they still
expect the initial URL to be queued as well. Fixes #330