Reformatted all source code with prettier. Changed indentation to 2 spaces.
I changed the indentation width because prettier places such heavy
emphasis on line length. With certain methods on the Crawler class,
the lines would be too heavily broken up with 4 spaces width
indentation. This change also aligns more closely with the JS standard
format (and is the default for prettier).