New documentation approach!
The README is now generated using jsdoc-to-markdown. This should
hopefully be helpful to us in a few different aspects. Obviously
we get the benefits of keeping code and documentation in the same
place. The JSDoc comments were already there, so this is just a way
to make use of them.

We also get the benefits of an simple documentation - keeping all
documentation in a single README is often appreciated for smaller
libraries, and it means we don't have to host a website.

jsdoc-to-markdown is also a real powerhouse, so it gives us a lot
of room for future improvements to the docs.